Resolved Errors in Generate
First error was in /posts. In the page/posts logic, I was returning the "page" obj. Nuxt doesn't like that for some reason. De-structured it with ...page and it worked out.

Second error was coming from nonexistent /home route. Storyblok doesn't let me change it to / so I did it in nuxt.config generate.